moved utility functions into own folder and created a package to use them
