+++ Project/utility/verification.py
@@ -1,5 +1,5 @@
 # Imports from other files
-from hash_util import hash_string_256, hash_block
+from utility.hash_util import hash_string_256, hash_block
 
 class Verification:
     @staticmethod
